1. update create bid busniess

+++ app/src/main/java/com/zhengdianfang/dazhongbao/models/product/Product.kt
@@ -10,7 +10,7 @@ import com.zhengdianfang.dazhongbao.models.login.User
 data class Product(var id: Long, var sharesCode: String, var sharesName: String, var yestodayClosePrice: Double,
                    var companyName: String?, var lastUnitPrice: Double, var nowUnitPrice: Double, var basicUnitPrice: Double,
                    var soldCount: Long, var limitTime: Int, var description: String?, var bidCount: Int, var check_status: Int,
-                   var contact: User?, var industry: String?, var attention: Int, var bond: Int, var bond_status: Int, var deal: MutableList<Deal>,
+                   var contact: User?, var industry: String?, var attention: Int, var bond: Double, var bond_status: Int, var deal: MutableList<Deal>,
                    var startDateTime: Long, var endDateTime: Long, var mybids: MutableList<Bid>?) {
 
     override fun equals(other: Any?): Boolean {
